weird, camera works on linux but not osx?
FossilOrigin-Name: 07fdca0dd33ea94398237fe631b09819856308a2b9a71a8ed6f5b0989524bc5f

+++ arcan_videoint.h
@@ -137,7 +137,7 @@ typedef struct arcan_vobject_litem arcan_vobject_litem;
 
 struct arcan_video_display {
 	bool suspended, text_support, fullscreen, conservative;
-	GLdouble projmatr[16];
+	GLfloat projmatr[16];
 	/* default image loading options */
 	enum arcan_vimage_mode scalemode;
 	GLuint deftxs, deftxt;
